mcp-external-server-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-external-server-python"
version = "0.1.0"
description = "Python version of mcp-external-server for testing mcp-ui-bridge-python"
authors = [
  { name="Your Name", email="you@example.com" }, # TODO: Replace with your actual name and email
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp-ui-bridge-python @ file:///E:/proyects/React%20to%20terminal%20conversor/mcp-ui-bridge-python",
    "typer>=0.9.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
]

[project.scripts]
start-server = "src.main:run_cli" # Assuming we'll wrap main.py's asyncio.run in a function for Typer if needed, or adjust if direct script execution is preferred. 